+++ syr2k_double_cublas.cu
@@ -18,14 +18,14 @@
 int main(int argc, char *argv[]) {
     cublasHandle_t cublasH = NULL;
 
-    int n = 16384, k = 16384;
+    long n = 16384, k = 16384;
 
     if (argc >= 3) {
         n = atoi(argv[1]);
         k = atoi(argv[2]);
     }
 
-    int lda = n, ldb = n, ldc = n;
+    long lda = n, ldb = n, ldc = n;
 
     double *d_A = nullptr;
     double *d_B = nullptr;
