+++ trsm_float_cublas.cu
@@ -13,14 +13,14 @@
 int main(int argc, char *argv[]) {
     cublasHandle_t cublasH = NULL;
 
-    int m = 16384, n = 16384;
+    long m = 16384, n = 16384;
 
     if (argc >= 5) {
         m = atoi(argv[1]);
         n = atoi(argv[2]);
     }
 
-    int lda = m, ldb = m;
+    long lda = m, ldb = m;
 
     float *d_A = nullptr;
     float *d_B = nullptr;
